EES-1816 Move ReleaseDataFileRepository to Content.Model
